chore: outsourced helper functions
